api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Denny Serejo <denny_serejo@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.4"}
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
httpx = "^0.27.0"
google-auth = "^2.32.0"
cryptography = "^43.0.0"
strawberry-graphql = "^0.237.3"
pytest-testmon = "^2.1.1"
pytest-watch = "^4.2.0"

[tool.poetry.group.dev.dependencies]
alembic = "^1.13.2"

[tool.poetry.group.test.dependencies]
pytest-mock = "^3.14.0"
pytest = "^8.3.2"
mock = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
